requirements.txt
asgiref==3.2.10
certifi==2020.11.8
chardet==3.0.4
Django==2.2
django-haystack==3.0
idna==2.10
mailchimp-marketing==3.0.27
Pillow==8.0.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.25.0
simplejson==3.17.2
six==1.15.0
spotipy==2.16.1
sqlparse==0.4.1
urllib3==1.26.2
Whoosh==2.7.4
